+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@ Release Notes
 =============
 ## 1.1.0-beta4
 * Cognitive Services: Retrieve ARM expression to the Key of the Cognitive Services instance.
+* Bastion Hosts: Create bastion hosts for accessing resources on a virtual network.
 
 ## 1.1.0-beta3
 * Web App / Functions: Allow CORS enable credentials (https://github.com/CompositionalIT/farmer/issues/265)